+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Better asserts on QKV dimensions [#264]
 - Better perfs for FusedMLP and FusedLinearLayer [#283]
 - Deepnorm init missing self-attention [#284]
+- Better self-attention projection [#299]
 
 ### Added
 - Simplicial Embeddings [#259]